examples/rfnry-chat/multi-tenant/client-python-b/pyproject.toml
[project]
name = "rfnry-chat-example-multi-tenant-agent-b"
version = "0.0.1"
description = "Multi-tenant example — standalone agent for organization-b (serves workspace-b-1 and workspace-b-2)"
requires-python = ">=3.11"
dependencies = [
    "rfnry-chat-protocol",
    "rfnry-chat-client",
    "anthropic>=0.80.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
rfnry-chat-protocol = { path = "../../../../../types/packages/chat-python", editable = true }
rfnry-chat-client = { path = "../../../../../chat/packages/client-python", editable = true }

[project.optional-dependencies]
dev = [
    "mypy>=1.20.1",
    "poethepoet>=0.44.0",
    "ruff>=0.15.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[[tool.mypy.overrides]]
module = ["socketio", "socketio.*"]
ignore_missing_imports = true

[tool.poe.tasks]
dev = "python -m src.main"
start = "python -m src.main"
format = "ruff format ."
"format:imports" = "ruff check --select I --fix ."
check = "ruff check ."
"check:fix" = "ruff check --fix ."
typecheck = "mypy src"
